Refacto UI : move code related to routing in router folder
