+++ ui/main/src/app/app.module.ts
@@ -14,7 +14,7 @@ import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {HttpClientModule, HTTP_INTERCEPTORS} from '@angular/common/http';
 import {CommonModule, HashLocationStrategy, LocationStrategy} from '@angular/common';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {AppRoutingModule} from './app-routing.module';
+import {AppRoutingModule} from './router/app-routing.module';
 import {AppComponent} from './app.component';
 import {LoginComponent} from './modules/core/application-loading/login/login.component';
 import {TranslateModule} from '@ngx-translate/core';
